+++ wal-line-flycheck.el
@@ -42,15 +42,31 @@
        'wal-line-neutral))
     (_ 'wal-line-neutral)))
 
+(defun wlf--get-error-help (status)
+  "Get the error count for STATUS."
+  (pcase status
+    ('running "Still checking")
+    ('finished
+     (if flycheck-current-errors
+         (let-alist (flycheck-count-errors flycheck-current-errors)
+           (format "Errors: %s, warnings: %s" (or .error 0) (or .warning 0)))
+       (format "Info: %s" (or .info 0))))
+    (_ "No problems")))
+
 (wal-line-create-augment flycheck
   :verify (lambda () (require 'flycheck nil t))
   :action
   (lambda (status &rest _r)
     (let ((face (wlf--get-face-for-status status))
+          (text (wlf--get-error-help status))
           (segment (wal-line-buffer-name--get-segment)))
-      (setq-local wal-line-buffer-name--segment (concat
-                                                 (wal-line--spacer)
-                                                 (propertize segment 'face face)))))
+      (setq-local wal-line-buffer-name--segment
+                  (concat
+                   (wal-line--spacer)
+                   (propertize
+                    segment
+                    'face face
+                    'help-echo text)))))
   :hooks
   (flycheck-status-changed-functions))